More conversion issues

+++ cocos2d/CCTexture2D.m
@@ -316,7 +316,7 @@ static CCTexture2DPixelFormat defaultAlphaPixel_format = kCCTexture2DPixelFormat
 			
 			NSUInteger neededBytes = (4 - mod ) / (bpp/8);
 
-			CCLOGWARN(@"cocos2d: WARNING converting size=(%d,%d) to size=(%d,%d) due to iOS 5.x memory BUG. See: http://www.cocos2d-iphone.org/forum/topic/31092", textureWidth, textureHeight, textureWidth + neededBytes, textureHeight );
+			CCLOGWARN(@"cocos2d: WARNING converting size=(%lu,%lu) to size=(%lu,%lu) due to iOS 5.x memory BUG. See: http://www.cocos2d-iphone.org/forum/topic/31092", (unsigned long)textureWidth, (unsigned long)textureHeight, (unsigned long)(textureWidth + neededBytes), (unsigned long)textureHeight );
 			textureWidth = textureWidth + neededBytes;
 		}
 	}   
@@ -647,9 +647,9 @@ static CCTexture2DPixelFormat defaultAlphaPixel_format = kCCTexture2DPixelFormat
 	else
 	{
 	#if CC_USE_LA88_LABELS
-		NSUInteger textureSize = textureWidth*textureHeight;
+		NSInteger textureSize = textureWidth*textureHeight;
 		unsigned short *la88_data = (unsigned short*)data;
-		for(int i = textureSize-1; i>=0; i--) //Convert A8 to AI88
+		for(NSInteger i = textureSize-1; i>=0; i--) //Convert A8 to AI88
 		la88_data[i] = (data[i] << 8) | 0xff;
 	#endif
 		 self = [self initWithData:data pixelFormat:LABEL_PIXEL_FORMAT pixelsWide:textureWidth pixelsHigh:textureHeight contentSize:computedDimension];
@@ -725,9 +725,9 @@ static CCTexture2DPixelFormat defaultAlphaPixel_format = kCCTexture2DPixelFormat
 	UIGraphicsPopContext();
 
 #if CC_USE_LA88_LABELS
-	NSUInteger textureSize = textureWidth*textureHeight;
+	NSInteger textureSize = textureWidth*textureHeight;
 	unsigned short *la88_data = (unsigned short*)data;
-	for(int i = textureSize-1; i>=0; i--) //Convert A8 to AI88
+	for(NSInteger i = textureSize-1; i>=0; i--) //Convert A8 to AI88
 		la88_data[i] = (data[i] << 8) | 0xff;
 
 #endif
